Create Irving's algorithm.py
This is an implementation of Irving's algorithm in Python. Based on the original algorithm, I add on two more features, one is to go through all the possible all-or-nothing cycles to find out all the solutions for a given problem, the other is to generate random samples and get the distribution of numbers of solutions.