Add files via upload
The stable roommate problem is a problem involving 2n people, trying to nd a
stable roommate assignment A that matches the 2n people into n disjoint pairs. A
roommate assignment A is called unstable if there is a blocking pair, such that each
of them prefers the other to his assigned roommate in A. An assignment A that
contains no blocking pair is called stable. An instance of a stable roommate problem
is called solvable if it has at least one stable assignment. An efficient algorithm to
detect whether a given instance of a stable roommate problem is solvable and to nd
one stable assignment if it exists, is first proposed by Irving (1985). The following of
this article would present an implementation of Irving's algorithm in Python for the
pedagogical purpose of facilitating interesting readers to understand this algorithm
and relevant issues about stable roommate problems.